Move all sodium tests into one tests file
